Rename qc_utils.ts
